Added S3 regression tests

+++ tests/test_MIRI.py
@@ -17,13 +17,14 @@ try:
     from eureka.S2_calibrations import s2_calibrate as s2
 except ModuleNotFoundError:
     pass
+
 from eureka.S3_data_reduction import s3_reduce as s3
 from eureka.S4_generate_lightcurves import s4_genLC as s4
 from eureka.S5_lightcurve_fitting import s5_fit as s5
 from eureka.S6_planet_spectra import s6_spectra as s6
 
 
-def test_MIRI(capsys):
+def test_MIRI(capsys, keep_s2_output, keep_s3_output):
     s2_installed = 'eureka.S2_calibrations.s2_calibrate' in sys.modules
     if not s2_installed:
         with capsys.disabled():
@@ -164,12 +165,13 @@ def test_MIRI(capsys):
     assert os.path.exists(name+os.sep+'figs')
 
     # remove temporary files
-    if s2_installed:
+    if s2_installed and not keep_s2_output:
         # os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}"
         #           f"Stage1{os.sep}S1_*")
         os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}"
                   f"Stage2{os.sep}S2_*")
-    os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage3")
+    if not keep_s3_output:
+        os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage3")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage4")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage5")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage6")
